# spec/openapi.yaml
openapi: 3.1.0
info:
  title: "clinic-ops-multistack API"
  version: "0.1.0"
  description: |
    動物病院の窓口業務システム。**この文書が唯一の正**（`spec/README.md`）。
    5つの実装（Go / Rails / Laravel / FastAPI / Next.js）は、この契約に対して外から
    同じ振る舞いをすること。中身（保存先・フレームワークの流儀）は自由。

    ## 画面のルートとデータのルート

    このシステムは基本的にサーバがHTMLを返す作り（Next.jsだけは事情が違うが、
    URLと振る舞いは揃える）。契約は2種類に分かれる。

    - **画面のルート**（`text/html` を返す）: 契約するのは **200が返ること** と
      **`x-data-testids` に列挙した目印が存在すること** の2点だけ。HTMLの見た目・
      マークアップの細部はスタックの流儀に任せる。
    - **データのルート**（`application/json` を返す／受け取る。`/dm.csv` のみ
      `text/csv`）: 値そのものを契約する。数値・件数・並び順の正しさは、
      **必ずこちら側で検証できるようにする。**

    ## HTMLフォーム送信時のエラーの出し方（重要）

    画面（HTML）ルートは、入力エラーや保存失敗が起きても **200 OK でフォームを
    再描画し**、打った値と一緒にエラー文言を `data-testid="error-banner"` に出す。
    ステータスコードで判定するのは常にデータのルート側であり、画面のルート側は
    「200 かつ error-banner の文言」で判定する。これは既存の実システムが
    実際に踏んだ事故（保存を断ったのに確定済みの値で描き直され、打った値が消えた）
    を避けるための形で、押し直しをさせないためにこの形にしてある。

    **文言そのものは画面・データの両ルートで一字一句共通。**
    下記「エラーの文言」の節にあるものを、状況に応じてそのまま使うこと。
    独自の文言を作らないこと（`coordination/DECISIONS.md` 第4節）。

    ## 日付・時刻

    **すべて JST（日本標準時、UTC+09:00）で扱う。** 集計の月境界・日境界もJST。
    `date-time` 型の値は `+09:00` オフセット付きのISO 8601で表す
    （例: `2026-09-05T10:30:00+09:00`）。サーバの内部時刻がUTCであっても、
    HTTPでやり取りする値はJSTに変換してから出し、受けるときもJSTとして解釈する。

    ## 文字数の数え方

    文字数の上限がある項目は、**Unicodeの文字数で数える。バイト数では数えない**
    （日本語はUTF-8で3倍になるため）。個別の上限値は `spec/acceptance.md` に従う。

    ## data-testid の命名規則

    共通テストがHTMLの中身を確認できるように、**目印は必ず `data-testid` 属性に置く**。

...

...
servers:
...
paths:
...
  /api/billings/{id}:
    get:
      operationId: "api_get_billing"
      tags: ["api-billing"]
      summary: "会計伝票1件の詳細（明細・合計・未算入件数を含む）"
      parameters:
        - $ref: "#/components/parameters/BillingId"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Billing" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: "api_update_billing"
      tags: ["api-billing"]
      summary: "会計伝票の更新（明細の追加・確定・支払い記録）"
      parameters:
        - $ref: "#/components/parameters/BillingId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BillingCreate" }
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Billing" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/ServerError" }
...
  /api/sales/summary:
    get:
      operationId: "api_sales_summary"
      tags: ["api-billing"]
      summary: "売上集計"
      description: |
        `Billing.status = confirmed` の伝票だけを対象にする。日・月の境界は
        **JST**。`unit_price` が未設定の明細行は集計から除外し、
        `excluded_detail_count_total` に総数を入れる。
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
        - name: group_by
          in: query
          required: false
          schema: { type: string, enum: ["day", "month", "category"], default: "day" }
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalesSummary" }
        "422": { $ref: "#/components/responses/ValidationError" }

  # ============================================================
  # データのルート — 入院・予約
  # ============================================================
  /api/ward:
    get:
      operationId: "api_ward_day"
      tags: ["api-ward"]
      summary: "指定日（既定はJSTの本日）に入院中の患者一覧"
      parameters:
        - name: date
          in: query
          required: false
          schema: { type: string, format: date }
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                required: ["items", "total"]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Hospitalization" }
                  total: { type: integer, minimum: 0 }
...
  /api/hospitalizations/{id}:
    get:
      operationId: "api_get_hospitalization"
      tags: ["api-ward"]
      summary: "入院1件の詳細（実施記録を含む）"
      parameters:
        - $ref: "#/components/parameters/HospitalizationId"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Hospitalization" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: "api_update_hospitalization"
      tags: ["api-ward"]
      summary: "入院情報の更新（退院日・処置室の変更を含む）"
      parameters:
        - $ref: "#/components/parameters/HospitalizationId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HospitalizationCreate" }
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Hospitalization" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/hospitalizations/{id}/care-records:
    get:
      operationId: "api_list_care_records"
      tags: ["api-ward"]
      summary: "実施記録の一覧（投薬・給餌・計測）"
      parameters:
        - $ref: "#/components/parameters/HospitalizationId"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                required: ["items", "total"]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/CareRecord" }
                  total: { type: integer, minimum: 0 }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      operationId: "api_create_care_record"
      tags: ["api-ward"]
      summary: "実施記録を1件追加"
      description: |
        **`staff_id`（実施者）は必須。** 空の記録行は作らない
        （`model.md` 15章「実施者が空の記録行を作らないこと」）。
        欠けている場合は 422 `invalid_input`。
      parameters:
        - $ref: "#/components/parameters/HospitalizationId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CareRecordCreate" }
      responses:
        "201":
          description: "作成できた"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CareRecord" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
...
  responses:
...
    BadRequest:
      description: "リクエストの本文がJSONとして壊れている。"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: "invalid_json"
              message: "リクエストの本文がJSONとして壊れています。書き方を確認してください。"
    ValidationError:
      description: "JSONとしては読めるが、項目が足りない・型が違う。"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: "invalid_input"
              message: "入力の形式が正しくありません。必須の項目や値の型を確認してください。"
              details:
                - field: "staff_id"
                  message: "実施者（staff_id）は必須です。"
    NotFound:
      description: "指定されたデータが見つからない。"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: "not_found"
              message: "指定されたデータが見つかりません。"
...
    ServerError:
      description: "保存に失敗した。"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: "save_failed"
              message: "保存に失敗しました。時間をおいてもう一度お試しください。"

  # ============================================================
  # 共通パラメータ
  # ============================================================
  parameters:
...
    BillingId:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
    HospitalizationId:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
  schemas:
    Error:
      type: object
      required: ["error"]
      properties:
        error:
          type: object
          required: ["code", "message"]
          properties:
            code:
              type: string
              enum:
                - "invalid_json"
                - "invalid_input"
                - "not_found"
                - "forbidden"
                - "save_failed"
                - "reservation_conflict"
            message:
              type: string
            details:
              type: array
              minItems: 1
              items:
                type: object
                required: ["field", "message"]
                properties:
                  field: { type: string }
                  message: { type: string }

    # ---- 1. Clinic ----
...
    BillingDetail:
      type: object
      required: ["row_no", "price_code", "name", "quantity", "is_taxable"]
      properties:
        id: { type: integer, readOnly: true }
        billing_id: { type: integer, readOnly: true }
        row_no: { type: integer, minimum: 1 }
        price_code: { type: string }
        name: { type: string }
        quantity: { type: number, exclusiveMinimum: 0 }
        unit_price:
          type: ["integer", "null"]
          description: "未設定（null）がありうる。**0円として合計に入れてはならない。**"
        is_taxable: { type: boolean }
        amount:
          type: ["integer", "null"]
          readOnly: true
          description: "quantity * unit_price。unit_price が null のときは null（合計に含めない）。"
    Billing:
      type: object
      required: ["id", "patient_id", "owner_id", "slip_no", "status", "billed_on", "details",
                 "net_amount", "tax_amount", "total_amount", "excluded_count"]
      properties:
        id: { type: integer, readOnly: true }
        patient_id: { type: integer }
        owner_id: { type: integer, readOnly: true }
        slip_no: { type: string, readOnly: true }
        status: { type: string, enum: ["draft", "confirmed"] }
        billed_on: { type: string, format: date }
        staff_id: { type: ["integer", "null"] }
        cashier_staff_id: { type: ["integer", "null"] }
        paid_amount: { type: ["integer", "null"] }
        payment_method: { type: ["string", "null"] }
        details:
          type: array
          items: { $ref: "#/components/schemas/BillingDetail" }
        net_amount:
          type: integer
          readOnly: true
          description: |
            **税抜合計。** unit_price が設定されている行だけを足したもの。
            未設定の行は0円として足さない（`spec/README.md`）。
            画面では `data-check="billing.net_amount"` に出す。
        taxable_subtotal:
          type: integer
          readOnly: true
          description: "そのうち課税対象の小計（税抜）。消費税はこれに税率を掛けて伝票単位で1回だけ切り捨てる。"
        nontaxable_subtotal:
          type: integer
          readOnly: true
          description: "unit_price が設定されている非課税行の小計。"
        tax_amount:
          type: integer
          readOnly: true
          description: "taxable_subtotal に Clinic.tax_rate を掛けた額。丸めの規則は `spec/acceptance.md` に従う（無指定なら丸めない）。"
        total_amount:
          type: integer
          readOnly: true
          description: "taxable_subtotal + tax_amount + nontaxable_subtotal。unit_price 未設定の行は含まない。"
        excluded_count:
          type: integer
          readOnly: true
          minimum: 0
          description: "unit_price が未設定で、total に含まれなかった明細行の数。0件でも必ず返す。"
    BillingCreate:
      type: object
      required: ["billed_on", "details"]
      properties:
        billed_on: { type: string, format: date }
        status: { type: string, enum: ["draft", "confirmed"], default: "draft" }
        staff_id: { type: ["integer", "null"] }
        cashier_staff_id: { type: ["integer", "null"] }
        paid_amount: { type: ["integer", "null"] }
        payment_method: { type: ["string", "null"] }
        details:
          type: array
          items:
            type: object
            required: ["price_code", "name", "quantity", "is_taxable"]
            properties:
              price_code: { type: string }
              name: { type: string }
              quantity: { type: number, exclusiveMinimum: 0 }
              unit_price: { type: ["integer", "null"] }
              is_taxable: { type: boolean }
...
    SalesSummary:
      type: object
      description: |
        売上を**3つの軸で同時に**返す。片方ずつ返す形（group_by）にしない。

        検算1は「分類別・担当別・日別の合計が、総合計と一致すること」を確かめるもので、
        **3つを同時に見られないと成立しない**（別々に取ると、取った時点がずれる余地も残る）。

        消費税は含めない。**3方向一致の対象は税抜売上のみ**。
        内訳へ税を按分すると、按分の方法しだいで答えが変わり検算にならない
        （`spec/acceptance.md` 検算1）。
      required: ["by_category", "by_staff", "by_date", "total_net_amount"]
      properties:
        from:
          type: ["string", "null"]
          format: date
          description: "省略可。省略したら全期間。"
        to:
          type: ["string", "null"]
          format: date
          description: "省略可。省略したら全期間。"
        by_category:
          type: array
          description: "分類別（料金の上位1階層）。"
          items: { $ref: "#/components/schemas/SalesRow" }
        by_staff:
          type: array
          description: "担当別。`Billing.staff_id`（診療担当）で分ける。`cashier_staff_id` は使わない。"
          items: { $ref: "#/components/schemas/SalesRow" }
        by_date:
          type: array
          description: "日別。`Billing.billed_on`（JSTの暦日）で分ける。"
          items: { $ref: "#/components/schemas/SalesRow" }
        total_net_amount:
          type: integer
          description: "対象期間の税抜売上の総合計。3つの内訳の合計はこれと一致すること。"
        total_tax_amount:
          type: integer
          description: "対象期間の全伝票の消費税額の合計。内訳へは按分しない。"
        excluded_count_total:
          type: integer
          description: "単価未設定で合計に入れなかった明細の行数。"
    SalesRow:
      type: object
      required: ["key", "net_amount"]
      properties:
        key:
          type: ["string", "integer"]
          description: "分類名・スタッフID・日付のいずれか。軸によって変わる。"
        label:
          type: string
          description: "画面に出す名前。"
        net_amount:
          type: integer
          description: "税抜。画面では `data-check=\"sales_summary.net_amount\"` に出す。"
        share_pct:
          type: number
          description: |
            構成比。**`by_category` の行にだけ付ける。**
            最大剰余法で丸め、**和がちょうど 100.0** になること。
            画面では `data-check="sales_summary.share_pct"`。
...
    CareRecord:
      type: object
      required: ["id", "hospitalization_id", "recorded_at", "category", "performed_by_staff_id"]
      properties:
        id: { type: integer, readOnly: true }
        hospitalization_id: { type: integer, readOnly: true }
        recorded_at: { type: string, format: date-time }
        category: { type: string, enum: ["medication", "feeding", "measurement"] }
        content: { type: ["string", "null"] }
        performed_by_staff_id:
          type: integer
          description: "実施者。**必須。** 空の記録行は作らない。"
    CareRecordCreate:
      type: object
      required: ["recorded_at", "category", "performed_by_staff_id"]
      properties:
        recorded_at: { type: string, format: date-time }
        category: { type: string, enum: ["medication", "feeding", "measurement"] }
        content: { type: ["string", "null"] }
        performed_by_staff_id: { type: integer }
    Hospitalization:
      type: object
      required: ["id", "patient_id", "admitted_on", "room"]
      properties:
        id: { type: integer, readOnly: true }
        patient_id: { type: integer }
        admitted_on: { type: string, format: date }
        discharged_on: { type: ["string", "null"], format: date }
        room: { type: string }
        care_records:
          type: array
          items: { $ref: "#/components/schemas/CareRecord" }
    HospitalizationCreate:
      type: object
      required: ["admitted_on", "room"]
      properties:
        admitted_on: { type: string, format: date }
        discharged_on: { type: ["string", "null"], format: date }
        room: { type: string }
...